added transcribe functions to utils.py then removed transribe from views.py of student and teacher
